+++ src/modules/palette/private/QuantizedPalette.cpp
@@ -1,0 +1,28 @@
+/**
+ * @file
+ */
+
+#include "QuantizedPalette.h"
+#include "core/Log.h"
+#include "core/RGBA.h"
+#include "core/collection/Buffer.h"
+
+namespace palette {
+
+bool QuantizedPalette::load(const core::String &filename, io::SeekableReadStream &stream, palette::Palette &palette) {
+	RGBAMap colors;
+	if (!load(filename, stream, colors)) {
+		return false;
+	}
+	const size_t colorCount = (int)colors.size();
+	core::Buffer<core::RGBA, 1024> colorBuffer;
+	colorBuffer.reserve(colorCount);
+	for (const auto &e : colors) {
+		colorBuffer.push_back(e->first);
+	}
+	palette.quantize(colorBuffer.data(), colorBuffer.size());
+	Log::debug("Loaded %i colors and quanitized to %i", (int)colorCount, palette.colorCount());
+	return palette.colorCount() > 0;
+}
+
+} // namespace palette