+++ docs/CHANGELOG.md
@@ -17,6 +17,7 @@ General:
    - Added new lua script to create newel stairs
    - Fixed missing animation transfer from one scenegraph into other one
    - Updated/Added SDL3 support (both SDL2 and SDL3 are supported now)
+   - Added Adobe ColorBook (`acb`) palette support
 
 VoxEdit: